fixed teamlist in teamOverview.jsp

+++ src/main/java/com/wemakeitwork/allenvooreen/model/Member.java
@@ -4,7 +4,15 @@ import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 
-import javax.persistence.*;
+import javax.persistence.CascadeType;
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.OneToMany;
+import javax.persistence.Table;
+import javax.persistence.Transient;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
@@ -16,7 +24,9 @@ public class Member implements UserDetails {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer memberId;
+    // Yes... this NEEDS to be a long, not an integer. Why? All lettuce you me dead:
+    // https://stackoverflow.com/questions/10997494/hibernate-error-ids-for-this-class-must-be-manually-assigned-before-calling-sav
+    private long memberId;
 
     private String memberName;
 
@@ -78,11 +88,11 @@ public class Member implements UserDetails {
         return true;
     }
 
-    public Integer getMemberId() {
+    public long getMemberId() {
         return memberId;
     }
 
-    public void setMemberId(Integer memberId) {
+    public void setMemberId(long memberId) {
         this.memberId = memberId;
     }
